chore: refactoring
Divided login page into reusable components: form inputs, sponsor info.
