--- src/pages/login.css
@@ -1,122 +1,0 @@
-.content-wrapper{
-    height: 100%;
-
-    display: grid;
-    gap: 50px;
-
-    grid-template-columns: 1fr 1fr;
-    justify-content: center;
-}
-
-.sponsor-info-wrapper {
-    display: flex;
-    align-items: center;
-    justify-content: flex-end;
-    min-height: 100%;
-}
-.sponsor-info {
-    width: 520px;
-    height: 605px;
-
-    h2{
-        text-align: center;
-        margin: 20px 0;
-    }
-
-    .image-holder.sponsor-image {
-        height: 270px;
-        width: 100%;
-        border-top-right-radius: 35px; 
-        border-top-left-radius: 35px; 
-    }
-
-    .credits-wrapper{
-        height: 272px;
-        background-color: var(--secondary-color);
-        border-radius: 0 0 35px 35px;
-
-        .credits{
-            padding-top: 70px;
-            padding-left: 65px;
-            color: var(--text-color);
-            font-weight: bold;
-
-            ul, li{
-                margin-left: 10px;
-            }
-        }
-    }
-}
-
-.input-form-wrapper {
-    display: flex;
-    align-items: center;
-    min-height: 100%;
-}
-.login-form-wrapper{
-
-    height: 530px;
-    width: 430px;
-
-    background-color: var(--base-color);
-
-    border-radius: 48px;
-
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: flex-end;
-
-    p.text-sm{
-        font-size:14px;
-        color: var(--accent-color);
-    }
-
-    .login-form{
-        display: flex;
-        flex-direction: column;
-        align-items: center;
-        margin-bottom: 40px;
-
-        h2{
-            margin-bottom: 30px;
-        }
-
-        form {
-            display: flex;
-            flex-direction: column;
-            align-items: center;
-
-            input{
-                width: 310px;
-                padding: 0 15px;
-                color: var(--accent-color);
-                font-weight: 500;
-
-                &::placeholder{
-                    color: var(--accent-color);
-                    font-weight: bold;
-                }
-
-                &:focus {
-                    border: 2px solid var(--secondary-color); 
-                    outline: none; 
-                }
-            }
-
-            .error-placeholder{
-                height:25px;
-            }
-        }
-
-        button{
-            margin-top: 25px;
-            margin-bottom: 35px;
-        }
-    }
-
-}
-
-.error-message{
-    color: var(--accent-color);
-}
