+++ src/pages/welcome/welcome-page.css
@@ -1,0 +1,62 @@
+.content-wrapper{
+    height: 100%;
+
+    display: grid;
+    gap: 50px;
+
+    grid-template-columns: 1fr 1fr;
+    justify-content: center;
+}
+
+.input-form-wrapper {
+    display: flex;
+    align-items: center;
+    min-height: 100%;
+}
+.form-wrapper{
+
+    height: 530px;
+    width: 430px;
+
+    background-color: var(--base-color);
+
+    border-radius: 48px;
+
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: flex-end;
+
+    p.text-sm{
+        font-size:14px;
+        color: var(--accent-color);
+    }
+
+    .form{
+        display: flex;
+        flex-direction: column;
+        align-items: center;
+        margin-bottom: 40px;
+
+        h2{
+            margin-bottom: 30px;
+        }
+
+        form {
+            display: flex;
+            flex-direction: column;
+            align-items: center;
+
+            input{
+                width: 310px;
+            }
+        }
+
+        button{
+            margin-top: 25px;
+            margin-bottom: 35px;
+        }
+    }
+
+}
+
